bacon: let the frameworks handle the camera.
Due of changes we have in fb this feature does not work well. We introduced
in the past a function in the keyguard which handles all possibilities which
can happen on the keyguard in combination with slim features.

On a later point we should do for a secure lockscreen if the camera
supports it a secure camera. Way it was I do not like due that to much
3rd party camera do not play well with the secure intent.

So like it is now is for now the best solution.

Change-Id: I15e092c2dcce1081e836c7e54760196c4936b2bf